Update TensorRT-LLM (#1455)

+++ cpp/tensorrt_llm/plugins/ncclPlugin/allreducePlugin.cpp
@@ -167,7 +167,7 @@ int AllreducePlugin::enqueue(nvinfer1::PluginTensorDesc const* inputDesc, nvinfe
     {
         return 0;
     }
-    int size = 1;
+    size_t size = 1;
     for (int i = 0; i < inputDesc[0].dims.nbDims; ++i)
     {
         size *= inputDesc[0].dims.d[i];
