+++ cpp/tensorrt_llm/plugins/ncclPlugin/recvPlugin.cpp
@@ -86,7 +86,7 @@ int RecvPlugin::enqueue(nvinfer1::PluginTensorDesc const* inputDesc, nvinfer1::P
     {
         return 0;
     }
-    int size = 1;
+    size_t size = 1;
     for (int i = 0; i < inputDesc[0].dims.nbDims; ++i)
     {
         size *= inputDesc[0].dims.d[i];
